requirements.txt
openai
pillow
pyperclip
python-dotenv
pynput